Remove abstract auth client class and simplify hierarchy
